libcrypto: Avoid incompatible pointers
Fixes: T164

+++ src/libcrypto.c
@@ -686,8 +686,12 @@ static int aes_ctr_set_key(struct ssh_cipher_struct *cipher, void *key,
     return SSH_OK;
 }
 
-static void aes_ctr_encrypt(struct ssh_cipher_struct *cipher, void *in, void *out,
-    unsigned long len) {
+static void
+aes_ctr_encrypt(struct ssh_cipher_struct *cipher,
+                void *in,
+                void *out,
+                size_t len)
+{
   unsigned char tmp_buffer[AES_BLOCK_SIZE];
   unsigned int num=0;
   /* Some things are special with ctr128 :
